requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-extensions==4.1
django-mathfilters==1.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docxtpl==0.20.2
Faker==40.1.0
gunicorn==25.1.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-docx==1.2.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
